+++ wx-cloud-common/src/main/java/com/tencent/wxcloudrun/constants/CategoryEnum.java
