+++ wx-cloud-common/src/main/java/com/tencent/wxcloudrun/request/WxPayCloseParam.java
@@ -1,4 +1,4 @@
-package com.tencent.wxcloudrun.model.request;
+package com.tencent.wxcloudrun.request;
 
 import lombok.Data;
 
@@ -9,6 +9,7 @@ import java.io.Serializable;
  * @date 2022/10/30
  */
 @Data
-public class WxPayQueryParam implements Serializable {
+public class WxPayCloseParam implements Serializable {
+
     private String out_trade_no;
 }
