+++ wx-cloud-common/src/main/java/com/tencent/wxcloudrun/response/Result.java
@@ -1,4 +1,4 @@
-package com.tencent.wxcloudrun.model.response;
+package com.tencent.wxcloudrun.response;
 
 import java.io.Serializable;
