+++ wx-cloud-web/src/main/java/com/tencent/wxcloudrun/controller/UserInfoController.java
@@ -5,10 +5,10 @@ package com.tencent.wxcloudrun.controller;
 import com.alibaba.fastjson.JSONObject;
 import com.tencent.wxcloudrun.annotation.ApiRequest;
 import com.tencent.wxcloudrun.entity.UserEntity;
-import com.tencent.wxcloudrun.model.request.UserCodeParam;
-import com.tencent.wxcloudrun.model.request.UserLoginParam;
-import com.tencent.wxcloudrun.model.request.UserInfoParam;
-import com.tencent.wxcloudrun.model.response.Result;
+import com.tencent.wxcloudrun.request.UserCodeParam;
+import com.tencent.wxcloudrun.request.UserInfoParam;
+import com.tencent.wxcloudrun.request.UserLoginParam;
+import com.tencent.wxcloudrun.response.Result;
 import com.tencent.wxcloudrun.service.UserInfoService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
