+++ wx-cloud-web/src/main/java/com/tencent/wxcloudrun/service/AdsInfoService.java
@@ -1,11 +1,9 @@
 package com.tencent.wxcloudrun.service;
 
 import com.alibaba.fastjson.JSONObject;
+import com.tencent.wxcloudrun.dto.PageDTO;
 import com.tencent.wxcloudrun.entity.AdsInfoEntity;
-import com.tencent.wxcloudrun.model.dto.PageDTO;
-import com.tencent.wxcloudrun.model.request.*;
-
-import java.util.List;
+import com.tencent.wxcloudrun.request.*;
 
 /**
  * @author tangsh
@@ -13,8 +11,6 @@ import java.util.List;
  */
 
 public interface AdsInfoService {
-
-    List<AdsInfoEntity> list();
 
     PageDTO<AdsInfoEntity> page(AdsPageParam param);
