+++ wx-cloud-web/src/main/java/com/tencent/wxcloudrun/service/impl/UserInfoServiceImpl.java
@@ -3,12 +3,12 @@ package com.tencent.wxcloudrun.service.impl;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.tencent.wxcloudrun.client.WxUserClient;
+import com.tencent.wxcloudrun.dto.RawDataDO;
 import com.tencent.wxcloudrun.entity.UserEntity;
-import com.tencent.wxcloudrun.model.dto.RawDataDO;
-import com.tencent.wxcloudrun.model.request.UserCodeParam;
-import com.tencent.wxcloudrun.model.request.UserInfoParam;
-import com.tencent.wxcloudrun.model.request.UserLoginParam;
 import com.tencent.wxcloudrun.repository.UserRepository;
+import com.tencent.wxcloudrun.request.UserCodeParam;
+import com.tencent.wxcloudrun.request.UserInfoParam;
+import com.tencent.wxcloudrun.request.UserLoginParam;
 import com.tencent.wxcloudrun.service.UserInfoService;
 import com.tencent.wxcloudrun.utils.EncryptUtils;
 import com.tencent.wxcloudrun.utils.NonceUtil;
